// ClientApp/src/modules/shared/interfaces/upperTask.interface.ts
export interface UpperTask
{
  workId: number;
  name: string;
  deliveryDate: Date;
  shopName: string;
  branchName: string;
  clothId: number;
  clothNames: string;
  height: number;
  front: number;
  collar: number;
  shoulder: number;
  sleeve: number;
  sleeveWidth: number;
  cuff: number;
  note: string;
  status: number;
  price: number;
  assignTo: string;
  paidTo: number;
  feedbackId: string;
}