// ClientApp/src/modules/shared/interfaces/lowerTask.interface.ts
export interface LowerTask
{
  workId: number;
  name: string;
  deliveryDate: Date;
  shopName: string;
  branchName: string;
  clothId: number;
  clothNames: string;
  height: number;
  knee: number;
  waist: number;
  seat: number;
  thigh: number;
  bottom: number;
  underline: number;
  note: string;
  status: number;
  price: number;
  assignTo: string;
  paidTo: number;
  feedbackId: string;
}